refactor(settings): split SettingsModal into one file per tab
The 602-line SettingsModal.tsx held nine components, well over the ~200-line
soft cap and the "one concern per file" rule in code-organization.md. Split it
into src/components/modals/settings/:

- controls.tsx — shared Toggle + Segmented primitives
- one file per tab (Appearance, Layout, Behavior, AI, Print, Privacy)
- SettingsModal.tsx — the shell (tab state, nav, footer)

No behavior change. Update the AppShell import and the App.test mock path, move
the modal test beside the new shell, and add tests that exercise every tab
(incl. AI and Print, which weren't rendered before) plus the custom-font and
AI-provider branches so per-file coverage holds.
