+++ src/components/modals/settings/controls.tsx
@@ -1,0 +1,42 @@
+// Shared form primitives used across the settings tabs.
+
+export function Toggle({
+  checked,
+  onChange,
+}: {
+  checked: boolean;
+  onChange: (v: boolean) => void;
+}) {
+  return (
+    <label className="settings-toggle">
+      <input type="checkbox" checked={checked} onChange={(e) => onChange(e.target.checked)} />
+      <span className="settings-toggle-track" />
+      <span className="settings-toggle-thumb" />
+    </label>
+  );
+}
+
+export function Segmented<T extends string>({
+  value,
+  options,
+  onChange,
+}: {
+  value: T;
+  options: { value: T; label: string }[];
+  onChange: (v: T) => void;
+}) {
+  return (
+    <div className="settings-segmented">
+      {options.map((opt) => (
+        <button
+          type="button"
+          key={opt.value}
+          data-active={value === opt.value}
+          onClick={() => onChange(opt.value)}
+        >
+          {opt.label}
+        </button>
+      ))}
+    </div>
+  );
+}
